invert {formula.tools}	R Documentation

invert

Description

Invert the operators in an object, usually a formula or expression

Usage

invert(x, ...)

## S4 method for signature 'call'
invert(x)

## S4 method for signature 'expression'
invert(x)

Arguments

x
function for invert

...
additional arguments passed other functions

invert is a S4 generic method for inverting relational operators, i.e.

functions prefixed with a . are not exported and should probably not be called directly

Value

The operand is returned with the relational operators inverted.

Author(s)

Christopher Brown

See Also

op, op.type

Examples

  invert( quote( A >  5 ) )
  invert( quote( A >= 5 ) )
  invert( quote( A <  5 ) )
  invert( quote( A <= 5 ) )
  invert( quote( A == 5 ) )
  invert( quote( A != 5 ) )
  invert( quote( A %in% lettters[1:5] ) )
  invert( quote( A %!in% letters[1:5] ) )

[Package formula.tools version 1.7.1 Index]