op.type {formula.tools}	R Documentation

Get the operator type used in an call, formula, expression, etc.

Description

Get the operator type used in an call, formula, expression, etc.

Usage

op.type(x)

## S4 method for signature 'call'
op.type(x)

## S4 method for signature 'formula'
op.type(x)

## S4 method for signature 'ANY'
op.type(x)

## S4 method for signature 'expression'
op.type(x)

## S4 method for signature 'list'
op.type(x)

Arguments

x
object from which to extract the operator type

Value

a character vector of the operator type(s)

See Also

op, operator.type

[Package formula.tools version 1.7.1 Index]