terms.call {formula.tools}	R Documentation

terms

Description

terms method for call and expression objects

Usage

## S3 method for class 'call'
terms(x, ...)

## S3 method for class 'expression'
terms(x, ...)

Arguments

x
A call object

...
Arguments passed to terms.formula

This S3 method returns a terms object for a call methods using a dispatch to terms.formula .

The terms are generated by making a rhs only call to terms.formula .

data is only needed and must be explicitly specified, i.e. data = if there are special elements such as '.'. Otherwise the data argument is unused.

Some edge cases may not be supported.

Value

A terms object. See terms.object for details.

Author(s)

Christopher Brown

See Also

stats::terms.object()

stats::terms.formula()

Examples

    terms( quote( A + B ) )

    data(iris) 
    x <- terms( quote( . - Species ) , data=iris ) 

[Package formula.tools version 1.7.1 Index]